+++ utils/obviosFunction.ts
@@ -1,0 +1,1 @@
+export const obviousFunction = (a: number, b: number) => a + b;
